refactor: Lotto와 common 폴더 생성
- Lotto 폴더에는 사용자 입력을 받아서 처리하는 파일들이 들어있음
- common 폴더에는 재사용 가능한 함수와 상수파일이 들어있음
